Better test. Better implementation
